// angular/src/app/shared/enums/date-range-selection-units.ts
//order in ascending only
export enum DateRangeUnits {
    CUSTOM = 'Selected interval',
    LAST_HOUR = 'Last hour',
    TODAY = 'Today',
    LAST_24_HOURS = 'Last 24 hours',
    THIS_WEEK = 'This week',
    LAST_7_DAYS = 'Last 7 days',
    THIS_MONTH = 'This month',
    LAST_30_DAYS = 'Last 30 days',
}